# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paytm Clone API",
    "version": "1.0.0",
    "description": "A node.js Express API"
  },
  "paths": {
...
    "/api/v1/getuserinfo/{id}": {
      "get": {
        "summary": "Fetch the information of the user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Id of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched the user info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully fetched the user info"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Fetching User."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sentfundsinfo/{id}": {
      "get": {
        "summary": "Fetch the history of sent funds",
        "tags": [
          "Transaction"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Id of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched the transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "senderName": {
                            "type": "string",
                            "example": "John Doe"
                          },
                          "receiverName": {
                            "type": "string",
                            "example": "Jane Smith"
                          },
                          "amountTransferred": {
                            "type": "string",
                            "example": "100.00"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully fetched the transactions."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No transaction found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No transaction found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching the transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error fetching transactions."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivedfundsinfo/{id}": {
      "get": {
        "summary": "Fetch the history of received funds",
        "tags": [
          "Transaction"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Id of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched the transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "senderName": {
                            "type": "string",
                            "example": "John Doe"
                          },
                          "receiverName": {
                            "type": "string",
                            "example": "Jane Smith"
                          },
                          "amountTransferred": {
                            "type": "string",
                            "example": "100.00"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully fetched the transactions."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No transaction found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No transaction found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching the transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error fetching transactions."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}